Move remaining vector molecules to be components in vector (for now)
