+++ src/components/views/rooms/SearchBar.js
